Move caching/loading methods to aikido::io
