+++ include/aikido/io.hpp
@@ -1,4 +1,5 @@
 #include "io/CatkinResourceRetriever.hpp"
 #include "io/KinBodyParser.hpp"
+#include "io/trajectory.hpp"
 #include "io/util.hpp"
 #include "io/yaml.hpp"
